% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_MultiLayer_Photosynthesis.R
\docType{data}
\name{FLAG}
\alias{FLAG}
\title{Func_Multi_Layer_Photosynthesis_Model}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
FLAG
}
\arguments{
\item{FLAG}{Model version controller;  0--Lloyd et al. 2010 Model for Vcmax-LAI relationship; 
1--Mercado et al. 2006 Model for Vcmax-LAI relationship in the tropics}

\item{SZA}{solar zenith angle, in degrees}

\item{Press}{Atmospheric pressure in Pa}

\item{PAR0}{measured  top canop irradiance, in umol/m2/s}

\item{LAI}{total canopy LAI}

\item{Tleaf}{Leaf temperature in degrees C}

\item{Tleaf_diff}{Difference in leaf temperature between sunlit and shaded leaf fractions 
in degrees C}

\item{ambCO2}{ambient CO2 concentration, in ppm}

\item{Vcmax0_25}{Vcmax at reference 25 degrees C for top-of-canopy leaves}

\item{CI}{Clumping inedx; 0.63 for tropical evergreen forests (Chen et al, 2005)}

\item{Topt}{optimal leaf temperature for tropical evergreen forests, from Lloyd and Farquhar, 2008}

\item{Nlayers}{number of layers for Multi-Layer Photosynthesis Modeling}

\item{Phi_sun}{the curvature factor for light response curves for sunlit leaves}

\item{PSII_sun}{maximum quantum yield for sunlit leaves}

\item{Phi_shade}{the curvature factor for light response curves for sunlit leaves}

\item{PSII_shade}{maximum quantum yield for shade leaves}

\item{sf_sun}{scaling factor for sunlit leaves, due to leaf age effect}

\item{sf_shade}{scaling factor for shade leaves, due to leaf age effect}

\item{sf}{scaling factor due to leaf age effect, assuming no phenological partitioning across vertical canopy profile}
}
\description{
Canopy multi-layer photosynthesis model
}
\author{
Jin Wu

Shawn Serbin
}
\keyword{datasets}
