% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Canopy_Radiation_Transfer.R
\name{Func_Canopy_Radiance_Transfer}
\alias{Func_Canopy_Radiance_Transfer}
\title{Func_Canopy_Radiance_Transfer}
\usage{
Func_Canopy_Radiance_Transfer(FLAG, SZA, LAI, Ib0, Id0, Vcmax0_25, CI)
}
\arguments{
\item{FLAG}{Model version controller;  0--Lloyd et al. 2010 Model for Vcmax-LAI relationship; 
1--Mercado et al. 2006 Model for Vcmax-LAI relationship in the tropics}

\item{SZA}{solar zenith angle, in degrees}
}
\description{
Function for revised DF1997 model to partition canopy LAI into sunlit/shade leaves LAI and
partition canopy Vcmax into sunlit/shade leaves Vcmax clumping index was added to original
DF1997 model, following the reference from Ryu et al. 2011

Inputs
}
\details{
Goals: using revised DF1997 model to partition canopy LAI into sunlit/shade leaves LAI and 
partition canopy Vcmax into sunlit/shade leaves Vcmax clumping index was added to original 
DF1997 model, following the reference from Ryu et al. 2011
}

