% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Light_Partitioning.R
\name{Func_Func_Light_Partitioning}
\alias{Func_Func_Light_Partitioning}
\title{Func_Light_Partitioning}
\usage{
Func_Func_Light_Partitioning(SZA, P, PAR)
}
\arguments{
\item{SZA}{solar zenith angle, in degrees}

\item{P}{Atmospheric Pressure, in pa}

\item{PAR}{measured total PAR?  umol/m2/s}
}
\value{
List containing: SZA - solar zenith angle, PAR - PAR, SV - total Visible light, 
SN - total NIR light, Ratio - the ratio between total measured light and total modeled light, 
fV - fraction of visble direct beam, fN - fraction of NIR direct beam, 
Model_DV - direct visible light, Model_dV - diffuse visible light, Model_DN - direct NIR light, 
Model_dN - diffuse NIR light
}
\description{
Function to partitioning incident radiatoin into direct and diffuse radiation, 
based on the Weiss and Norman, 1985 light partitioning approach
}
\details{
Weiss and Norman, 1985 light partitioning approach
}
\author{
Jin Wu

Shawn Serbin
}
\references{
Weiss and Norman, 1985
}

