% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Leaf_FvCB_Model.R
\name{Func_Leaf_FvCB_Photosynthesis_Model}
\alias{Func_Leaf_FvCB_Photosynthesis_Model}
\title{Func_Leaf_FvCB_Photosynthesis_Model}
\usage{
Func_Leaf_FvCB_Photosynthesis_Model(Vcmax25, Jmax25, Tleaf, Topt, I, Ci, Press,
  PSII_in, Phi_in)
}
\arguments{
\item{Vcmax25}{Vcmax at 25 degrees C}

\item{Jmax25}{Jmax at 25 degrees C}

\item{Tleaf}{Leaf temperature}

\item{Topt}{Temperature optimum for Jmax}

\item{I}{incident light}

\item{Ci}{internal CO2 concentration in umols/mol or ppm}

\item{Press}{Atmospheric pressure in Pa}

\item{PSII_in}{Input PSII for maximum quantum yield}

\item{Phi_in}{Input curvature factor for light response function}
}
\description{
Leaf level FvCB Photosynthesis model (Farquhar et al. 1980)
}

