% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temperature_Response_Functions.R
\name{Func_Temperature_Bernacchi}
\alias{Func_Temperature_Bernacchi}
\title{Func_Temperature_Bernacchi}
\usage{
Func_Temperature_Bernacchi(delta_H, c, T)
}
\arguments{
\item{delta_H}{Activation energy}

\item{c}{Scaling constant}

\item{T}{leaf temperature in degrees C}
}
\value{
temperature scale factor
}
\description{
Bernacchi temperature response function for photosynthesis parameters
}
\details{
Bernacchi temperature response function for photosynthesis parameters
}
\author{
Jin Wu

Shawn Serbin
}

