% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temperature_Response_Functions.R
\name{Func_Temperature_June}
\alias{Func_Temperature_June}
\title{Func_Temperature_June}
\usage{
Func_Temperature_June(P25, Topt, Tleaf)
}
\arguments{
\item{P25}{Vcmax/Jmax at 25 degrees C}

\item{Topt}{Temperature optimum of Vcmax/Jmax}

\item{Tleaf}{leaf temperature in degrees C}
}
\description{
June temperature response function for photosynthesis parameters
}
\details{
June temperature response function for photosynthesis parameters
}
\references{
Bernacchi et al. 2013 and June et al. 2004
}
\author{
Jin Wu

Shawn Serbin
}
