% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temperature_Response_Functions.R
\name{Func_Temperature_Medlyn}
\alias{Func_Temperature_Medlyn}
\title{Func_Temperature_Medlyn}
\usage{
Func_Temperature_Medlyn(Tleaf.1, Tleaf.2, Param, Ea)
}
\arguments{
\item{Tleaf.1}{Original leaf temperature (degrees C)}

\item{Tleaf.2}{Leaf temperature to scale parameter to (degrees C)}

\item{Param}{Parameter (e.g. Vcmax, Kc, Gamma*) value at Tleaf.1}

\item{Ea}{Activation energy for Arrehnius function.  Dependent on parameter of interest}
}
\value{
Parameter scaled to Tleaf.2 following the Arrhenius function in Medlyn et al., 2002
}
\description{
Medlyn Arrhenius function
}
\details{
Medlyn Arrhenius temperature response function for photosynthesis parameters, based
on Bernacchi et al. 2002.  Used for Vcmax, Jmax, Rd, Kc, Ko, Gamma*
}
\references{
Bernacchi et al., 2002; 2003; Medlyn et al., 2002
}
\author{
Jin Wu

Shawn Serbin
}
