% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temperature_Response_Functions.R
\name{Func_Temperature_Response}
\alias{Func_Temperature_Response}
\title{Func_Temperature_Response}
\usage{
Func_Temperature_Response(V25, J25, Tleaf, Topt, Press)
}
\arguments{
\item{J25}{Jmax at 25 degrees C}

\item{Tleaf}{leaf temperature in degrees C}

\item{Topt}{Temperature optimum of Vcmax/Jmax}

\item{Press}{Atmospheric pressure in Pa}

\item{VC25}{Vcmax at 25 degrees C}
}
\value{
List containing: Vcmax - Vcmax at leaf temperature, Jmax - Jmax at leaf temperature,
Gamma_star - Gamma* at leaf temperature, Kc - Kc MM constant at leaf temperature, Ko - Ko MM constant at leaf temperature,
PSII - PSII at leaf temperature, Phi - Phi at leaf temperature, Rd - leaf respiration at leaf temperature,
Vomax - Vomax (max oxygen evolution) at leaf temperature
}
\description{
Temperature response functions for scaling leaf-level photosynthesis parameters
}
\details{
Temperature response functions for scaling leaf-level photosynthesis parameters
}
\references{
Bernacchi et al., 2002, 2003, 2013
}
\author{
Jin Wu

Shawn Serbin
}
