% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{model.options}
\alias{model.options}
\title{parse model.options.xml file used to set parameters and other options for model runs}
\usage{
model.options(input.file = NULL)
}
\arguments{
\item{input.file}{model.options.xml file containing information needed for run}
}
\description{
Read model.options.xml file
}
\examples{
\dontrun{
opt <- model.options()
model.options <- model.options('/home/$USER/model.options.xml')
}

}
\author{
Jin Wu, Shawn P. Serbin
}
