% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Light_Partitioning.R
\name{Func_Func_Light_Partitioning}
\alias{Func_Func_Light_Partitioning}
\title{Func_Light_Partitioning}
\usage{
Func_Func_Light_Partitioning(SZA, P, PAR)
}
\arguments{
\item{SZA}{solar zenith angle, in degrees}

\item{P}{air pressure in ??}

\item{PAR}{measured total PAR?}
}
\value{
List containing: SZA - solar zenith angle, PAR - PAR, SV - 

 @references Weiss and Norman, 1985
}
\description{
Function to partitioning incident radiatoin into direct and diffuse radiation, 
based on the Weiss and Norman, 1985 light partitioning approach
}
\details{
Weiss and Norman, 1985 light partitioning approach
}
\author{
Jin Wu

Shawn Serbin
}

