% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Leaf_FvCB_Model.R
\name{Func_Leaf_FvCB_Photosynthesis_Model}
\alias{Func_Leaf_FvCB_Photosynthesis_Model}
\title{Func_Leaf_FvCB_Photosynthesis_Model}
\usage{
Func_Leaf_FvCB_Photosynthesis_Model(
  Vcmax25,
  Jmax25,
  Tleaf,
  Topt,
  I,
  Ci,
  Press,
  PSII_in,
  Phi_in
)
}
\arguments{
\item{Vcmax25}{Vcmax at 25 degrees C}

\item{Jmax25}{Jmax at 25 degrees C}

\item{Tleaf}{Leaf temperature in degrees C}

\item{Topt}{Temperature optimum for Jmax}

\item{I}{incident light}

\item{Ci}{internal CO2 concentration in umols/mol or ppm}

\item{Press}{Atmospheric pressure in Pa}

\item{PSII_in}{Input PSII for maximum quantum yield}

\item{Phi_in}{Input curvature factor for light response function}
}
\value{
List containing: Anet - Net photosynthetic rate (Agross-Rd,umol/m2/s) at leaf temperature, 
Agross - Gross photosynthetic rate (umol/m2/s) at leaf temperature, Rd - Leaf respiration rate (umol/m2/s) at leaf temperature, 
Wc - Rubisco limited photosynthetic rate (umol/m2/s), Wj - RuBP limited photosynthetic rate (umol/m2/s),
Wp - TPU limited photosynthetic rate (0.5*Vcmax, umol/m2/s), Vcmax - Maximum rate of RuBP carboxylation (umol/m2/s) 
at leaf temperature, Jmax - Maximum rate of electron transport (umol/m2/s) at leaf temperature,
Vo - , Vomax - Maximum rate of oxygen evolution (umol/m2/s) at leaf temperature, Gamma* - at leaf temperature,
Kc - , Ko - , PSII - , Phi - at leaf temperature
}
\description{
Leaf level FvCB Photosynthesis model (Farquhar et al. 1980)
}
\examples{
\dontrun{
# Run test leaf photosynthesis simulations
file <- system.file('examples/Example_Leaf_Photosynthesis.R',package='CanopyPhotosynthesis')
source(file)

}

}
\references{
Long and Bernacchi, 2003; Medlyn et al., 2002; Bernacchi et al., 2013
}
\author{
Jin Wu

Shawn Serbin
}
